Prepare sign and ecp verify sign for soap.
Added functionality to prepare signature for a authn request, without
signing. This makes it possible to sign the authn request within a SOAP
message instead. To verify a authn request in a ecp it have to validate
the whole SOAP message and not just the authn request. Added logic to
sen the original xml document to the validation function.