Implement initiating logout from s2repoze plugin.
Adds a new repoze config option path_logout to the challenge decider
plugin. When a request is received on this url a global logout request
is initiated. Responses to this request is received on the
single_logout_service endpoint of the sp as configured in saml config
of the sp. Tested with only on IdP and only using HTTP_REDIRECT bindings

TODO: handle receiving logout requests on the single_logout_service endpoint