Change examples to bind to 127.0.0.1 by default.
Since the README already says to point a browser at localhost, it
was surprising to see the listeners bound to 0.0.0.0 by default.
Changed that and added notes in README about how to change it if
the user really wants to make a test listener accessible
externally.

Updated output messages during startup to include the bound IP,
to make this more obvious.

Added a missing EOL or two.