Fixed problem with metadata extensions
With this small modification, I fixed I problem I was experiencing while trying to add RequestInitiator extension to a SP metadata.
The problem here is that the XML attributes are called Binding and Location while the python instance attributes are called binding and location respectively.
Se the instruction "elif key in _inst.c_attributes:" needs to have key with the capital letter while the setattr needs the lowercase name of the field. To be more generic I imagined it was better to retrieve the field name from the tuple associated to the field in c_attributes.