Remove tests asserting unnecessary py2 rounding
f_quotient in time_utils does not actually ever get passed values in the
ranges that are being tested here. It is, however, asserting bad python2
rounding behavior in the int() constructor that has been removed. We can
just remove these tests, and be comfortable reasoning that there aren't
time differences that cause this situation based on the docs of
time.struct_time.