Remove signature verification directly on origdoc
The old logic always tried to verify signature directly on the (encrypted) origdoc
(original message), which often led to unnecessary xmlsec1 failures, accompanied
by an excessive and misleading logs.

The code such as this should never have to "guess" the content and act upon its
"best belief" as what the content actually is (or looks like). This "too smart"
behavior is commonly considered dangerous and has no place in security libraries.