Fix Name-ID caching type inconsistency
If we already promoted (in #0515de9f) the saml2.cache.Cache to be a bit smart when it
comes to understanding the data it stores - that is decoding name_id in the get() method,
let's be consistent and do the reverse operation directly in the set() method.

This also fixes the issue with djangosaml2 that PR #321 tried to solve (but ugly way).