The ID of each Signature element must be unique
If the assertion and response both are signed, both Signatures have an
ID of `Signature1`. This creates invalid xml as xs:ID must be unique.
This fixes the issue when integrating with onelogin's python3-saml
client:
Element '{http://www.w3.org/2000/09/xmldsig#}Signature', attribute 'Id':
'Signature1' is not a valid value of the atomic type 'xs:ID'.