Fix SSRF caused by URI attribute of Reference element
The Signature element contains a Reference element that has a URI
attribute that points to the data object that has been signed. The URI
attribute is of type anyURI which means it can have a value of a local
or remote resource. This can cause system leaks, exceptions or cross
site attacks.

To fix this, xmlsec1 is limited to alow empty an empty URI or resolve
same-doc URI schemes. Local or external URIs will be rejected.

The problem is also described in the XML Signature Best Practices
documents:
https://www.w3.org/TR/xmldsig-bestpractices/#problematic-external-references
https://www.w3.org/2008/xmlsec/Drafts/best-practices/Overview.html#problematic-external-references

Fixes #510 Thanks @marcpare

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>