Update cache to use the datetime module
This will be a recurring theme in the commits that follow. The check can be
explained as follows:

the original check is:

    if ... time_util.after(timestamp): raise ...

which can be thought of as:

    is_vaild = not time_util.after(datetime)

means that this is valid if the current datetime is not after the given
datetime. IOW, this is valid if the current datetime is before the given
datetime or, swapping the comparison elements, this is valid if the given
datetime is after the current datetime.

The equivalent new check is:

    is_valid = saml2.datetime.compare.after_now(date_time)

means that this is valid if the given datetime is after the current datetime.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>