Multiple AttributeStatement tags per Assertion
This was necessary to implement a real-world SSO integration,
which required handlinge multiple AttributeStatement elements
within a single assertion in a SAML response.

Orginally this change was implemented in a private fork
by Thomas Knott for pysaml 2.2.0, and has been ported
by Sheila Allen for use in pysaml 4.6.0 to hopefully
merge upstream.

There was a similar PR for the same need by pcrownov:
https://github.com/IdentityPython/pysaml2/pull/205/files