Allow any namespace prefix
If the xsd string comes with a namespace prefix, it is left intact.
If the xsd string has no namespace prefix and the xsd type is known,
the xsd namespace is set to 'xs'.
If the xsd string has no namespace and the type is unknown,
the xsd namespace is left empty.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>