Do not hardcode the warning filter
The application should control whether warnings should be visible or not. By
hardcoding the simplefilter we turn on warnings' visibility for all modules
that follow. Removing this allows the application code to decide if warnings
should be shown.

To enable warnings through the command line pass -Wd to the python interpreter.

Quoting the python warnings module documentation[0]:
> You can do this from the command-line by passing -Wd to the interpreter (this
> is shorthand for -W default). This enables default handling for all warnings,
> including those that are ignored by default. To change what action is taken
> for encountered warnings you simply change what argument is passed to -W,
> e.g. -W error. See the -W flag for more details on what is possible.

  [0]: https://docs.python.org/2/library/warnings.html#updating-code-for-new-versions-of-python

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>