Update saml_uri.py
According to https://tools.ietf.org/html/rfc2985 the urn for emailAddress has to be `1.2.840.113549.1.9.1`.
In saml_uri.py this is not implemented correctly. The current version uses `PKCS_9+'1'` which equals to `1.2.840.113549.1.9.1.1`. This can be fixed by deleting the trailing '1.' from line 8. This should not cause any side-effects, because the variable `PKCS_9` in combination with `+'1'`.