Make AESCipher and Fernet compatible
The two schemes now have compatible methods and signatures. To do this, Fernet accepts
more params than it uses. These have been added to allow for easier transition. They
will be removed in the next release to clean up the class.

Moreover, Fernet dictates a very specific format for the "key" that it accepts an uses
to encrypt and decrypt payloads. The key must be 32 url-safe bytes long in base64
format.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>