Correctly sign an AuthnRequest with Redirect binding
When an AuthnRequest is created with HTTP-Redirect binding, the XML document is not
signed, but instead, a signature is calculated and becomes part of the query params of
the Redirect-URL, through the Signature and SignAlg params.

Previously, when the Redirect binding was requested and signing was enabled but no
SignAlg params were defined, the Signature and SignAlg query params would be missing.

Now, if no SignAlg is defined, the default is used and the request is correctly created
with the proper query params.

Signed-off-by: Ivan Kanakarakis <ivan.kanak@gmail.com>