fix: mdstore: fix exception handler in InMemoryMetaData.parse
The exception handler in InMemoryMetaData.parse was failing for subclasses
other then `MetaDataFile` with:

    AttributeError: 'MetaDataExtern' object has no attribute 'filename'

- because `self.filename` is only defined for MetaDataFile but not MetaDataExtern

The handler was essentially expecting it would only be invoked for MetaDataFile
and not other subclasses of InMemoryMetaData.

Provide useful descriptive messages for MetaDataFile and MetaDataExtern
subclassses - and fall back to a generic (but safe) message otherwise.