new: saml2.Entity: support reloading metadata
Support reloading metadata by adding a reload_metadata method to saml2.Entity.

This method gets the metadata configuration in the same format
as the 'metadata' entry in the configuration passed to saml2.Config.

To keep metadata refreshed, this method needs to be periodically explicitly called.

For a metadata refresh with the same configuration, the calling application
should keep a copy of the original configuration to pass to this method.

Resolves #808