Reload metadata in-place
Metadata reloading was previously implemented by loading the metadata,
then replacing references to the old metadata with the new metadata. A
bug in the implementation caused the previous version of the metadata to
be indirectly referenced by the new version of the metadata, resulting
in a steady climb in memory usage.

In fixing the memory leak, I have also changed how metadata is reloaded
to avoid having to replace all existing references, which is prone to
errors and could cause confusing behaviour.